Did a whole work around the /myplaytime command. Reworked the method so that the user gets responded with message if there are no records of games in the server they are in, and also validate and check if the user write a gamename where there are no records of it in the server.
So the lists just display the games they have played, and if they have records from other games, but write manually on a new game in the server where they have no records, the bot will respond with that there are no records for that specific game.

The command also respond to the user if there are no games recorded at all in that server.