Did a lot of rework to the project.
Changed the tracking main on per server so it tracks user when the user has joined voice, so it makes a record based on the server they are on, and not updating each server where the bot and same user is on. So a valid tracking time of games happens when user is on voice in the particular server now.

A lot of small changes, cleaning up code, and last debugging of database, since the flushing didn't went into the database, but now the liveminutes are logged.